requirements.txt
groq
requests
python-dotenv
rich
